requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.3.8
attrs==25.3.0
bcrypt==4.3.0
black==23.9.1
blacken-docs==1.14.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
classify-imports==4.2.0
click==8.1.8
codespell==2.2.6
colorama==0.4.6
colored==2.3.0
colorlog==6.9.0
commonmark==0.9.1
contourpy==1.3.2
coverage==7.6.12
cycler==0.12.1
dill==0.3.9
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-stubs==5.2.0
django-stubs-ext==5.2.0
djangorestframework==3.15.2
djangorestframework-stubs==3.15.3
djangorestframework_simplejwt==5.5.0
docutils==0.21.2
docx==0.2.4
dotenv==0.9.9
drf-yasg==1.21.9
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==37.0.0
filelock==3.18.0
flake8==6.1.0
flake8-blind-except==0.2.1
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-docstrings==1.7.0
flake8-implicit-str-concat==0.5.0
fonttools==4.57.0
gunicorn==23.0.0
identify==2.6.10
idna==3.10
imageio==2.37.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
joblib==1.5.0
jsondiff==2.2.1
kiwisolver==1.4.8
lazy_loader==0.4
lxml==5.3.1
matplotlib==3.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.5
packaging==24.2
PassportEye==2.2.2
pathspec==0.12.1
pdfminer==20191125
phonenumbers==9.0.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.2.0
psycopg2-binary==2.9.10
py-unused-deps==0.4.2
pycodestyle==2.11.1
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pydocstyle==6.3.0
pyflakes==3.1.0
Pygments==2.19.1
PyJWT==2.9.0
pylibmc==1.6.3
pylint==3.3.4
pymemcache==4.0.0
pypandoc==1.15
pyparsing==3.2.3
pytesseract==0.3.13
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-memcached==1.62
pytz==2025.1
pyupgrade==3.9.0
PyYAML==6.0.2
reorder-python-imports==3.12.0
requests==2.32.3
rich==12.6.0
rstcheck==6.1.2
rstcheck-core==1.2.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
simple_zpl2==0.3.0
six==1.17.0
snowballstemmer==2.2.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenize_rt==6.1.0
tomli==2.2.1
tomlkit==0.13.2
typer==0.7.0
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
virtualenv==20.30.0
